moved tools to common
