+++ tools/src/common/bitmap.cpp
