+++ tools/src/common/bitmap.h
