+++ tools/src/common/palette.cpp
