+++ tools/src/common/palette.h
